Refactor(hooks): add useCountdown to barrel file and import it
